requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
playwright>=1.40.0
flask>=3.0.0
gunicorn>=22.0.0
python-telegram-bot>=21.0
